requirements.txt
argparse
requests
azure-devops
pandas
numpy
matplotlib